+++ packages/solid/CHANGELOG.md
@@ -3,10 +3,12 @@
 ### 💥 BREAKING CHANGES
 
 - `CheckboxControl` and `CheckboxLabel` components has been removed.
+- `SwitchControl` and `SwitchLabel` components has been removed.
 
 ### ✨ Features
 
 - Added `CheckboxPrimitive` and `CheckboxPrimitiveIndicator` unstyled component.
+- Added `SwitchPrimitive` and `SwitchPrimitiveThumb` unstyled component.
 
 ## [0.4.2](https://github.com/fabien-ml/hope-ui/compare/v0.4.1...v0.4.2) (2022-04-08)